% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{betaDeltaDiscountFunc}
\alias{betaDeltaDiscountFunc}
\title{Beta Delta Value Function}
\usage{
betaDeltaDiscountFunc(x, beta, delta)
}
\arguments{
\item{x}{observation at point n (X)}

\item{beta}{fitted parameter}

\item{delta}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Beta Delta Value Function
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
