% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{betaDeltaDiscountGradient}
\alias{betaDeltaDiscountGradient}
\title{Beta Delta Gradient Helper for Nonlinear Fitting}
\usage{
betaDeltaDiscountGradient(x, beta, delta)
}
\arguments{
\item{x}{observation at point n (X)}

\item{beta}{fitted parameter}

\item{delta}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Beta Delta Gradient Helper for Nonlinear Fitting
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
