% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{BleichrodtCRDIDiscountGradient}
\alias{BleichrodtCRDIDiscountGradient}
\title{Bleichrodt et al. Constant Relative Decreasing Impatience (CRDI) Helper for Nonlinear Fitting}
\usage{
BleichrodtCRDIDiscountGradient(x, lnk, s, beta)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}

\item{s}{fitted parameter}

\item{beta}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Bleichrodt et al. Constant Relative Decreasing Impatience (CRDI) Helper for Nonlinear Fitting
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
