% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounting.R
\name{dd_analyze}
\alias{dd_analyze}
\title{dd_analyze}
\usage{
dd_analyze(fittingObject)
}
\arguments{
\item{fittingObject}{core dd fitting object}
}
\value{

}
\description{
this is the business end of the analytical process
}
