% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleichrodt.R
\name{dd_discount_func_bleichrodt_crdi}
\alias{dd_discount_func_bleichrodt_crdi}
\title{Bleichrodt et al. Constant Relative Decreasing Impatience (CRDI) Value Function}
\usage{
dd_discount_func_bleichrodt_crdi(x, lnk, s, beta)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}

\item{s}{fitted parameter}

\item{beta}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Bleichrodt et al. Constant Relative Decreasing Impatience (CRDI) Value Function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
