% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_ebertprelec.R
\name{dd_discount_func_ebertprelec}
\alias{dd_discount_func_ebertprelec}
\title{Ebert & Prelec Value Function}
\usage{
dd_discount_func_ebertprelec(x, lnk, s)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}

\item{s}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Ebert & Prelec Value Function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
