% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_exponential.R
\name{dd_discount_func_exponential}
\alias{dd_discount_func_exponential}
\title{Exponential discounting function}
\usage{
dd_discount_func_exponential(x, lnk)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Exponential discounting function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
