% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_greenmyerson.R
\name{dd_discount_func_greenmyerson}
\alias{dd_discount_func_greenmyerson}
\title{Green & Myerson Value Function}
\usage{
dd_discount_func_greenmyerson(x, lnk, s)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}

\item{s}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Green & Myerson Value Function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
