% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_laibson.R
\name{dd_discount_func_laibson}
\alias{dd_discount_func_laibson}
\title{Beta Delta Value Function}
\usage{
dd_discount_func_laibson(x, beta, delta)
}
\arguments{
\item{x}{observation at point n (X)}

\item{beta}{fitted parameter}

\item{delta}{fitted parameter}
}
\description{
Beta Delta Value Function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
