% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_mazur.R
\name{dd_discount_func_mazur}
\alias{dd_discount_func_mazur}
\title{Hyperbolic Value Function}
\usage{
dd_discount_func_mazur(x, lnk)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Hyperbolic Value Function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
