% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_rachlin.R
\name{dd_discount_func_rachlin}
\alias{dd_discount_func_rachlin}
\title{Rachlin Value Function}
\usage{
dd_discount_func_rachlin(x, lnk, s)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}

\item{s}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Rachlin Value Function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
