% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mazur.R
\name{dd_discount_grad_mazur}
\alias{dd_discount_grad_mazur}
\title{Hyperbolic Gradient Helper for Nonlinear Fitting}
\usage{
dd_discount_grad_mazur(x, lnk)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Hyperbolic Gradient Helper for Nonlinear Fitting
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
