% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rachlin.R
\name{dd_discount_grad_rachlin}
\alias{dd_discount_grad_rachlin}
\title{Rachlin Gradient Helper for Nonlinear Fitting}
\usage{
dd_discount_grad_rachlin(x, lnk, s)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}

\item{s}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Rachlin Gradient Helper for Nonlinear Fitting
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
