% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_rodriguezlogue.R
\name{dd_discount_grad_rodriguezlogue}
\alias{dd_discount_grad_rodriguezlogue}
\title{Rodriguez & Logue Helper for Nonlinear Fitting}
\usage{
dd_discount_grad_rodriguezlogue(x, lnk, beta)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}

\item{beta}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Rodriguez & Logue Helper for Nonlinear Fitting
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
