% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleichrodt.R
\name{dd_ed50_bleichrodt}
\alias{dd_ed50_bleichrodt}
\title{dd_ed50_bleichrodt}
\usage{
dd_ed50_bleichrodt(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\description{
dd_ed50_bleichrodt
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
