% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_exponential.R
\name{dd_ed50_exponential}
\alias{dd_ed50_exponential}
\title{dd_ed50_exponential}
\usage{
dd_ed50_exponential(Lnk)
}
\arguments{
\item{Lnk}{log transformed rate parameter}
}
\description{
dd_ed50_exponential
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
