% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_greenmyerson.R
\name{dd_ed50_greenmyerson}
\alias{dd_ed50_greenmyerson}
\title{dd_ed50_greenmyerson}
\usage{
dd_ed50_greenmyerson(Lnk, s)
}
\arguments{
\item{Lnk}{parameter}

\item{s}{parameter}
}
\description{
dd_ed50_greenmyerson
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
