% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_laibson.R
\name{dd_ed50_laibson}
\alias{dd_ed50_laibson}
\title{dd_ed50_laibson}
\usage{
dd_ed50_laibson(b, d)
}
\arguments{
\item{b}{beta param}

\item{d}{delta param}
}
\description{
dd_ed50_laibson
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
