% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mazur.R
\name{dd_ed50_mazur}
\alias{dd_ed50_mazur}
\title{dd_ed50_mazur}
\usage{
dd_ed50_mazur(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{

}
\description{
dd_ed50_mazur
}
