% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_rodriguezlogue.R
\name{dd_ed50_rodriguezlogue}
\alias{dd_ed50_rodriguezlogue}
\title{dd_ed50_rodriguezlogue}
\usage{
dd_ed50_rodriguezlogue(Lnk, b, currentData)
}
\arguments{
\item{Lnk}{parameter}

\item{b}{parameter}

\item{currentData}{current data}
}
\description{
dd_ed50_rodriguezlogue
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
