% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{dd_fit_noise}
\alias{dd_fit_noise}
\title{dd_fit_noise}
\usage{
dd_fit_noise(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{

}
\description{
This fits an intercept only model to the data. Its trash, but its a testable alternative that inferring usefulness from an R2 value
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
