% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rachlin.R
\name{dd_fit_rachlin}
\alias{dd_fit_rachlin}
\title{dd_fit_rachlin}
\usage{
dd_fit_rachlin(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{

}
\description{
This fits a hyperbolic model to the data.
}
