% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_laibson.R
\name{dd_integrand_laibson_log10}
\alias{dd_integrand_laibson_log10}
\title{Beta Delta Integrand helper (log10)}
\usage{
dd_integrand_laibson_log10(x, beta, delta)
}
\arguments{
\item{x}{observation at point n (X)}

\item{beta}{fitted parameter}

\item{delta}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
This integrand helper is a projection of the integrand with delays represented in the log base 10 scale
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
