% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_rachlin.R
\name{dd_integrand_rachlin_log10}
\alias{dd_integrand_rachlin_log10}
\title{Rachlin Integrand helper (log10)}
\usage{
dd_integrand_rachlin_log10(x, lnK, s)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnK}{fitted parameter}

\item{s}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
This integrand helper is a projection of the integrand with delays represented in the log base 10 scale
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
