% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodriguezlogue.R
\name{dd_integrand_rodriguezlogue}
\alias{dd_integrand_rodriguezlogue}
\title{Rodriguez & Logue Integrand helper}
\usage{
dd_integrand_rodriguezlogue(x, lnK, beta)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnK}{fitted parameter}

\item{beta}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
This integrand helper is a projection of the integrand with delays represented as normal
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
