% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_rodriguezlogue.R
\name{dd_integrand_rodriguezlogue_log10}
\alias{dd_integrand_rodriguezlogue_log10}
\title{Rodriguez & Logue Integrand helper}
\usage{
dd_integrand_rodriguezlogue_log10(x, lnK, beta)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnK}{fitted parameter}

\item{beta}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
This integrand helper is a projection of the integrand (log10) with delays represented in the log base 10 scale
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
