% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebertprelec.R
\name{dd_mbauc_ebertprelec}
\alias{dd_mbauc_ebertprelec}
\title{dd_mbauc_ebertprelec}
\usage{
dd_mbauc_ebertprelec(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{

}
\description{
dd_mbauc_ebertprelec
}
