% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{dd_mbauc_exponential}
\alias{dd_mbauc_exponential}
\title{dd_mbauc_exponential}
\usage{
dd_mbauc_exponential(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{

}
\description{
dd_mbauc_exponential
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
