% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_laibson.R
\name{dd_mbauc_laibson}
\alias{dd_mbauc_laibson}
\title{dd_mbauc_laibson}
\usage{
dd_mbauc_laibson(A, b, d, startDelay, endDelay)
}
\arguments{
\item{A}{maximum value}

\item{b}{parameter value}

\item{d}{parameter value}

\item{startDelay}{time point}

\item{endDelay}{time point}
}
\description{
dd_mbauc_laibson
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
