% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_bleichrodt.R
\name{dd_mbauc_log10_bleichrodt}
\alias{dd_mbauc_log10_bleichrodt}
\title{dd_mbauc_log10_bleichrodt}
\usage{
dd_mbauc_log10_bleichrodt(A, Lnk, s, b, startDelay, endDelay)
}
\arguments{
\item{A}{maximum value}

\item{Lnk}{parameter value}

\item{s}{parameter value}

\item{b}{parameter value}

\item{startDelay}{time point}

\item{endDelay}{time point}
}
\description{
dd_mbauc_log10_bleichrodt
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
