% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebertprelec.R
\name{dd_mbauc_log10_ebertprelec}
\alias{dd_mbauc_log10_ebertprelec}
\title{dd_mbauc_log10_ebertprelec}
\usage{
dd_mbauc_log10_ebertprelec(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\description{
dd_mbauc_log10_ebertprelec
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
