% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_greenmyerson.R
\name{dd_mbauc_log10_greenmyerson}
\alias{dd_mbauc_log10_greenmyerson}
\title{dd_mbauc_log10_greenmyerson}
\usage{
dd_mbauc_log10_greenmyerson(A, Lnk, s, startDelay, endDelay)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\description{
dd_mbauc_log10_greenmyerson
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
