% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_mazur.R
\name{dd_mbauc_log10_mazur}
\alias{dd_mbauc_log10_mazur}
\title{dd_mbauc_log10_mazur}
\usage{
dd_mbauc_log10_mazur(A, Lnk, startDelay, endDelay)
}
\arguments{
\item{A}{maximum value of good}

\item{Lnk}{log transformed rate parameter}

\item{startDelay}{start delay}

\item{endDelay}{end delay}
}
\description{
dd_mbauc_log10_mazur
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
