% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{dd_mbauc_log10_noise}
\alias{dd_mbauc_log10_noise}
\title{dd_mbauc_log10_noise}
\usage{
dd_mbauc_log10_noise(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{

}
\description{
dd_mbauc_log10_noise
}
