% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mazur.R
\name{dd_mbauc_mazur_exact}
\alias{dd_mbauc_mazur_exact}
\title{dd_mbauc_mazur_exact}
\usage{
dd_mbauc_mazur_exact(A, hypLnk, startDelay, endDelay)
}
\arguments{
\item{A}{maximum value}

\item{hypLnk}{logged parameter value}

\item{startDelay}{time point}

\item{endDelay}{time point}
}
\value{

}
\description{
dd_mbauc_mazur_exact
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
