% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_rachlin.R
\name{dd_mbauc_rachlin}
\alias{dd_mbauc_rachlin}
\title{dd_mbauc_rachlin}
\usage{
dd_mbauc_rachlin(A, Lnk, s, startDelay, endDelay)
}
\arguments{
\item{A}{maximum value}

\item{Lnk}{parameter value}

\item{s}{parameter value}

\item{startDelay}{time point}

\item{endDelay}{time point}
}
\description{
dd_mbauc_rachlin
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
