% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodriguezlogue.R
\name{dd_mbauc_rodriguezlogue}
\alias{dd_mbauc_rodriguezlogue}
\title{dd_mbauc_rodriguezlogue}
\usage{
dd_mbauc_rodriguezlogue(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\description{
dd_mbauc_rodriguezlogue
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
