% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_call_metric_options.R
\name{dd_metric_options}
\alias{dd_metric_options}
\title{dd_metric_options}
\usage{
dd_metric_options(fittingObject, metrics)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{metrics}{(char vector) This vector contains a list of possible cross-model metrics.}
}
\value{

}
\description{
This call builds out the metrics used to conduct cross-model comparisons (e.g., ED50, MBAUC)
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
