% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounting.R
\name{dd_metricOptions}
\alias{dd_metricOptions}
\title{dd_metricOptions}
\usage{
dd_metricOptions(fittingObject, metrics)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{metrics}{vector specifying metrics}
}
\value{

}
\description{
method to indicate metrics of interest in the analysis
}
