% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_call_dd_model_options.R
\name{dd_model_options}
\alias{dd_model_options}
\title{dd_model_options}
\usage{
dd_model_options(fittingObject, plan)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{plan}{(char vector) This vector contains a list of possible model candidates.}
}
\value{

}
\description{
This call builds out the model(s) included in subsequent regressions.
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
