% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounting.R
\name{dd_modelOptions}
\alias{dd_modelOptions}
\title{dd_modelOptions}
\usage{
dd_modelOptions(fittingObject, plan)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{plan}{vector of model candidates}
}
\value{

}
\description{
method to specify which models to include as candidates
}
