% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_probable_model.R
\name{dd_probable_model}
\alias{dd_probable_model}
\title{dd_probable_model}
\usage{
dd_probable_model(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\description{
This method is used to perform approximate Bayesian model selection using extracted Bayes Factors from calculated BIC values.
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
