% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounting.R
\name{dd_probableModel}
\alias{dd_probableModel}
\title{dd_probableModel}
\usage{
dd_probableModel(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{

}
\description{
This method is used to perform approximate Bayesian model selection using extracted Bayes Factors from calculated BIC values.
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
