% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebertprelec.R
\name{dd_start_ebertprelec}
\alias{dd_start_ebertprelec}
\title{dd_start_ebertprelec}
\usage{
dd_start_ebertprelec(currentData)
}
\arguments{
\item{currentData}{current data set}
}
\value{

}
\description{
Extract starting parameters
}
