% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_exponential.R
\name{dd_start_exponential}
\alias{dd_start_exponential}
\title{dd_start_exponential}
\usage{
dd_start_exponential(currentData, increment = 1)
}
\arguments{
\item{currentData}{current data set}

\item{increment}{step size for span}
}
\description{
Extract starting parameters
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
