% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rachlin.R
\name{dd_start_rachlin}
\alias{dd_start_rachlin}
\title{dd_start_rachlin}
\usage{
dd_start_rachlin(currentData)
}
\arguments{
\item{currentData}{current data set}
}
\description{
Extract starting parameters
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
