% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{discountingModelSelection}
\alias{discountingModelSelection}
\title{Perform Discounting Model Selection}
\usage{
discountingModelSelection(dat, models = c("noise"), figures = FALSE,
  summarize = FALSE, lineSize = 1)
}
\arguments{
\item{dat}{data frame with X column and Y column (0 <= Y <= 1)}

\item{models}{vector of models to include in selection}

\item{figures}{OPTIONAL: show figure of results}

\item{summarize}{OPTIONAL: Descriptive observations}

\item{lineSize}{OPTIONAL: Modify line sizes for figures}
}
\value{
A data frame of model parameters

data frame of fitted model parameters
}
\description{
This function takes a data frame of temporal discounting values (X, Y)
and performs approximate Bayesian model selection using the Bayesian
Information Criterion and returns the log of the Effective Delay
50 and numerical integration area.
}
\details{
Models: Exponential, Hyperbolic, BetaDelta, GreenMyerson, & Rachlin, Ebert & Prelec's Constant Sensitivity
}
\examples{
discountingModelSelection(data.frame(X=c(1,30,180,540,1080,2160),
Y=c(1.0,0.9,0.8,0.7,0.6,0.4)),
models=c("noise", "exponential", "hyperbolic", "bd", "mg", "rachlin", "cs",
Figures = FALSE))
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
