% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{discountingModelSelectionCall}
\alias{discountingModelSelectionCall}
\title{Perform Discounting Model Selection}
\usage{
discountingModelSelectionCall(dat, A = NULL, models = c("noise"),
  detailed = FALSE, figures = FALSE, summarize = FALSE, lineSize = 1)
}
\arguments{
\item{dat}{data frame with X column and Y column (0 <= Y <= 1)}

\item{A}{OPTIONAL: Modify line sizes for figures}

\item{models}{vector of models to include in selection}

\item{detailed}{include additional output details in results}

\item{figures}{OPTIONAL: show figure of results}

\item{summarize}{OPTIONAL: Descriptive observations}

\item{lineSize}{OPTIONAL: Modify line sizes for figures}
}
\value{
A data frame of model parameters

data frame of fitted model parameters
}
\description{
This function takes a data frame of temporal discounting values (X, Y) and performs approximate Bayesian model selection using the Bayesian Information Criterion and returns the log of the Effective Delay 50 and numerical integration area.
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
