% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{displayAUCFigure}
\alias{displayAUCFigure}
\title{Display most probable fitted series with AUC metric}
\usage{
displayAUCFigure(dat, results, lineWidth = 1)
}
\arguments{
\item{dat}{observed data}

\item{results}{Results of analyses for data series}

\item{lineWidth}{Line width}
}
\value{
display figure
}
\description{
Display most probable fitted series with AUC metric
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
