% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{displayED50Figure}
\alias{displayED50Figure}
\title{Display of all fitted series with ED50 metric}
\usage{
displayED50Figure(dat, results, lineWidth = 1)
}
\arguments{
\item{dat}{observed data}

\item{results}{Results of analyses for data series}

\item{lineWidth}{Line width}
}
\value{
display figure
}
\description{
Display of all fitted series with ED50 metric
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
