% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{displayLogAUCFigure}
\alias{displayLogAUCFigure}
\title{Display most probable fitted series with AUC metric (log10)}
\usage{
displayLogAUCFigure(dat, results, lineWidth = 1)
}
\arguments{
\item{dat}{observed data}

\item{results}{Results of analyses for data series}

\item{lineWidth}{Line width}
}
\value{
display figure
}
\description{
Display most probable fitted series with AUC metric (log10)
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
