% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebertprelec.R
\name{ebertPrelecDiscountFunc}
\alias{ebertPrelecDiscountFunc}
\title{Ebert & Prelec Value Function}
\usage{
ebertPrelecDiscountFunc(x, lnk, s)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}

\item{s}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Ebert & Prelec Value Function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
