% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ebertPrelecDiscountGradient}
\alias{ebertPrelecDiscountGradient}
\title{Ebert & Prelec Helper for Nonlinear Fitting}
\usage{
ebertPrelecDiscountGradient(x, lnk, s)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}

\item{s}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Ebert & Prelec Helper for Nonlinear Fitting
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
