% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{exponentialDiscountFunc}
\alias{exponentialDiscountFunc}
\title{Exponential Value Function}
\usage{
exponentialDiscountFunc(x, lnk)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Exponential Value Function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
