% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounting.R
\name{fitDDCurves}
\alias{fitDDCurves}
\title{fitDDCurves}
\usage{
fitDDCurves(data, settings, maxValue)
}
\arguments{
\item{data}{assigned data}

\item{settings}{mappings}

\item{maxValue}{A parameter}
}
\value{

}
\description{
Core fitting object
}
