% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{gauss_2F1}
\alias{gauss_2F1}
\title{Workaround for varying bx for hypergeometric series}
\usage{
gauss_2F1(a, b, c, x)
}
\arguments{
\item{a}{param}

\item{b}{param}

\item{c}{param}

\item{x}{param}
}
\description{
Credit: Stéphane Laurent <laurent_step at yahoo.fr>
Source: https://stats.stackexchange.com/questions/33451/computation-of-hypergeometric-function-in-r
Licensed CC-BY-SA 3.0, as Per SA Guidelines
}
\author{
Stéphane Laurent <laurent_step at yahoo.fr>
}
