% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_metric_get_ED50.R
\name{get_ED50}
\alias{get_ED50}
\title{get_ED50}
\usage{
get_ED50(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{
natural logarithm of the Effective Delay 50%
}
\description{
Scoring for the log ED50
}
\details{
Methods for routing computations of the ln(ED50). When possible, exact solutions are provided and models without a straightforward exact solution have ed50 calculated numerically using a bisection search.
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
