% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_metric_get_MBAUC.R
\name{get_MBAUC}
\alias{get_MBAUC}
\title{Scoring for the most probable model area}
\usage{
get_MBAUC(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{
area beneath the fitted model
}
\description{
In this set of methods, the area beneath the fitted model is calculated and divided by the maximum area using numerical integration methods.  All delays are calculated in the normal scale.
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
