% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getED50}
\alias{getED50}
\title{Scoring for the log ED50}
\usage{
getED50(dat, results)
}
\arguments{
\item{dat}{observed data}

\item{results}{Results of analyses for data series}
}
\value{
natural logarithm of the Effective Delay 50%
}
\description{
Optionally, models without a straightforward exact solution
will be scored numerically using a bisection search
}
\details{
Only Ebert & Prelec, 2007 requires this bisection search
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
