% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getED50CS}
\alias{getED50CS}
\title{Numerically solve for ED50 value for Ebert & Prelec}
\usage{
getED50CS(dat, results)
}
\arguments{
\item{dat}{observed data}

\item{results}{Results of analyses for data series}
}
\value{
effective delay (value) for Ebert & Prelec CS
}
\description{
Numerically solve for ED50 value for Ebert & Prelec
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
