% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaymetrics.R
\name{getED50genhyp}
\alias{getED50genhyp}
\title{Numerically solve for ED50 value for Rodriguez & Logue Generalized Hyperbolioid}
\usage{
getED50genhyp(dat, results)
}
\arguments{
\item{dat}{observed data}

\item{results}{Results of analyses for data series}
}
\value{
effective delay (value) for Ebert & Prelec ep
}
\description{
This method solves for ED50 for Rodriguez & Logue using a point bisection procedure. This procedure will continue for n (20 currently) by default until a value of 50% is observed in the midpoint of two more moving delays.
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
