% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounting.R
\name{getMBAUCLog10}
\alias{getMBAUCLog10}
\title{Scoring for the most probable model area, in log10 space}
\usage{
getMBAUCLog10(dat, results)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{
area beneath the fitted model, in log10 space
}
\description{
In this set of methods, the area beneath the fitted model is calculated and divided by the maximum area using numerical integration methods.  All delays are calculated in the log base 10 scale.
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
