% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getModelAUC}
\alias{getModelAUC}
\title{Scoring for the most probable model area}
\usage{
getModelAUC(dat, results)
}
\arguments{
\item{dat}{observed data}

\item{results}{Results of analyses for data series}
}
\value{
area beneath the fitted model
}
\description{
Scoring for the most probable model area
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
