% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mazur.R
\name{hyperbolicDiscountFunc}
\alias{hyperbolicDiscountFunc}
\title{Hyperbolic Value Function}
\usage{
hyperbolicDiscountFunc(x, lnk)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Hyperbolic Value Function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
