% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{hyperbolicDiscountGradient}
\alias{hyperbolicDiscountGradient}
\title{Hyperbolic Gradient Helper for Nonlinear Fitting}
\usage{
hyperbolicDiscountGradient(x, lnk)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Hyperbolic Gradient Helper for Nonlinear Fitting
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
