% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_bleichrodt.R
\name{integrand_bleichrodt_crdi_log10}
\alias{integrand_bleichrodt_crdi_log10}
\title{Bleichrodt et al. Constant Relative Decreasing Impatience (CRDI) Integrand helper (log10)}
\usage{
integrand_bleichrodt_crdi_log10(x, lnK, s, beta)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnK}{fitted parameter}

\item{s}{fitted parameter}

\item{beta}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
This integrand helper is a projection of the integrand with delays represented in the log base 10 scale
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
