% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_ebertprelec.R
\name{integrand_ebertprelec_log10}
\alias{integrand_ebertprelec_log10}
\title{Ebert & Prelec's ep Integrand helper (log10)}
\usage{
integrand_ebertprelec_log10(x, lnK, s)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnK}{fitted parameter}

\item{s}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
This integrand helper is a projection of the integrand with delays represented in the log base 10 scale
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
