% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{integrandBetaDelta}
\alias{integrandBetaDelta}
\title{Beta Delta Integrand helper}
\usage{
integrandBetaDelta(x, beta, delta)
}
\arguments{
\item{x}{observation at point n (X)}

\item{beta}{fitted parameter}

\item{delta}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
Beta Delta Integrand helper
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
