% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleichrodt.R
\name{integrandBleichrodtCRDI}
\alias{integrandBleichrodtCRDI}
\title{Bleichrodt et al. Constant Relative Decreasing Impatience (CRDI) Integrand helper}
\usage{
integrandBleichrodtCRDI(x, lnK, s, beta)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnK}{fitted parameter}

\item{s}{fitted parameter}

\item{beta}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
This integrand helper is a projection of the integrand with delays represented as normal
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
