% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{integrandExp}
\alias{integrandExp}
\title{Exponential Integrand helper}
\usage{
integrandExp(x, lnK)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnK}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
This integrand helper is a projection of the integrand
with delays represented as normal
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
