% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{integrandHyp}
\alias{integrandHyp}
\title{Hyperbolic Integrand helper}
\usage{
integrandHyp(x, lnK)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnK}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
Hyperbolic Integrand helper
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
