% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrandHypLog}
\alias{integrandHypLog}
\title{Hyperbolic Integrand helper (log10)}
\usage{
integrandHypLog(x, lnK)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnK}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
This integrand helper is a projection of the integrand with delays represented in the log base 10 scale
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
