% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{integrandMyerson}
\alias{integrandMyerson}
\title{Green & Myerson Integrand helper}
\usage{
integrandMyerson(x, lnK, s)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnK}{fitted parameter}

\item{s}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
Green & Myerson Integrand helper
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
