% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{integrandMyersonLog}
\alias{integrandMyersonLog}
\title{Green & Myerson Integrand helper (log10)}
\usage{
integrandMyersonLog(x, lnK, s)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnK}{fitted parameter}

\item{s}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
Green & Myerson Integrand helper (log10)
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
