% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrandRachlin}
\alias{integrandRachlin}
\title{Rachlin Integrand helper}
\usage{
integrandRachlin(x, lnK, s)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnK}{fitted parameter}

\item{s}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
This integrand helper is a projection of the integrand with delays represented as normal
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
