% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrandRodriguezLogue}
\alias{integrandRodriguezLogue}
\title{Rodriguez & Logue Integrand helper}
\usage{
integrandRodriguezLogue(x, lnK, beta)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnK}{fitted parameter}

\item{beta}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
This integrand helper is a projection of the integrand
with delays represented as normal
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
