% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrandRodriguezLogueLog}
\alias{integrandRodriguezLogueLog}
\title{Rodriguez & Logue Integrand helper}
\usage{
integrandRodriguezLogueLog(x, lnK, beta)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnK}{fitted parameter}

\item{beta}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
This integrand helper is a projection of the integrand (log10) with delays represented in the log base 10 scale
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
