% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{logLik.nls.lm}
\alias{logLik.nls.lm}
\title{minpack.lm logLik hack}
\usage{
\method{logLik}{nls.lm}(fit, REML = FALSE, ...)
}
\arguments{
\item{fit}{nls.lm object}

\item{REML}{REML}

\item{...}{inherit}
}
\value{
provide a logLik class for AIC/BIC
}
\description{
minpack.lm logLik hack
}
\author{
Katharine Mullen <kate@few.vu.nl>
}
