% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_message_debug.R
\name{message_debug}
\alias{message_debug}
\title{message_debug}
\usage{
message_debug(fittingObject, msg)
}
\arguments{
\item{fittingObject}{core fitting object}

\item{msg}{(char) message}
}
\value{

}
\description{
Extension of message method, instead yolked to a flag defining level of verbosity.
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
