% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{messageDebug}
\alias{messageDebug}
\title{messageDebug}
\usage{
messageDebug(fittingObject, msg)
}
\arguments{
\item{fittingObject}{core fitting object}

\item{msg}{message}
}
\value{

}
\description{
messageDebug
}
