% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenmyerson.R
\name{myersonHyperboloidDiscountFunc}
\alias{myersonHyperboloidDiscountFunc}
\title{Green & Myerson Value Function}
\usage{
myersonHyperboloidDiscountFunc(x, lnk, s)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}

\item{s}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Green & Myerson Value Function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
