% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenmyerson.R
\name{myersonHyperboloidDiscountGradient}
\alias{myersonHyperboloidDiscountGradient}
\title{Green & Myerson Gradient Helper for Nonlinear Fitting}
\usage{
myersonHyperboloidDiscountGradient(x, lnk, s)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}

\item{s}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Green & Myerson Gradient Helper for Nonlinear Fitting
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
