% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.discountingtools}
\alias{plot.discountingtools}
\title{plot.discountingtools}
\usage{
\method{plot}{discountingtools}(
  fittingObject,
  position0 = "bottomleft",
  ylab0 = "Subjective Value",
  xlab0 = "Delay",
  logAxis = "x",
  yMin = 0.01
)
}
\arguments{
\item{fittingObject}{core frame}

\item{position0}{position legend}

\item{ylab0}{y axis label}

\item{xlab0}{x axis label}

\item{logAxis}{axis designation}

\item{yMin}{y axis lower limit}
}
\value{

}
\description{
Override plot output
}
