% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_plot_cross_model.R
\name{plot_cross_rainbow}
\alias{plot_cross_rainbow}
\title{plot_cross_rainbow}
\usage{
plot_cross_rainbow(fittingObject, metric, plotit)
}
\arguments{
\item{fittingObject}{core fitting object}

\item{metric}{(char) the cross model metric to be displayed}

\item{plotit}{(logical) bool of whether or not to print visual or output plotting frame}
}
\description{
plot_cross_rainbow
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
