% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_plot_group.R
\name{plot_group_rainbow}
\alias{plot_group_rainbow}
\title{plot_group_rainbow}
\usage{
plot_group_rainbow(
  fittingObject,
  position0,
  ylab0,
  xlab0,
  logAxis,
  yMin,
  plotit
)
}
\arguments{
\item{fittingObject}{core fitting object}

\item{position0}{(char) position of legend}

\item{ylab0}{(char) y axis label}

\item{xlab0}{(char) x axis label}

\item{logAxis}{(char) axis designation}

\item{yMin}{(num) y axis lower limit}

\item{plotit}{(logical) bool of whether or not to print visual or output plotting frame}
}
\description{
Convenience method for illustrating individual fits when characterized by some a priori grouping.
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
