% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_plot_model.R
\name{plot_model_characterization}
\alias{plot_model_characterization}
\title{plot_model_characterization}
\usage{
plot_model_characterization(fittingObject, position0, ylab0, xlab0, plotit)
}
\arguments{
\item{fittingObject}{core fitting object}

\item{position0}{(char) position of legend}

\item{ylab0}{(char) y axis label}

\item{xlab0}{(char) x axis label}

\item{plotit}{(logical) bool of whether or not to print visual or output plotting frame}
}
\description{
plot_model_characterization
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
