% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotGroupRainbow}
\alias{plotGroupRainbow}
\title{plotGroupRainbow}
\usage{
plotGroupRainbow(fittingObject, position0, ylab0, xlab0, logAxis, yMin)
}
\arguments{
\item{fittingObject}{core frame}

\item{position0}{position legend}

\item{ylab0}{y axis label}

\item{xlab0}{x axis label}

\item{logAxis}{axis designation}

\item{yMin}{y axis lower limit}
}
\value{

}
\description{
plotGroupRainbow
}
