% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_plot_ind_detailed.R
\name{plotIndividualDetailed}
\alias{plotIndividualDetailed}
\title{plotIndividualDetailed}
\usage{
plotIndividualDetailed(
  fittingObject,
  position0,
  ylab0,
  xlab0,
  logAxis,
  yMin,
  id,
  plotit
)
}
\arguments{
\item{fittingObject}{core fitting object}

\item{position0}{(char) position of legend}

\item{ylab0}{(char) y axis label}

\item{xlab0}{(char) x axis label}

\item{logAxis}{(char) axis designation}

\item{yMin}{(num) y axis lower limit}

\item{id}{(num) participant id}

\item{plotit}{(logical) bool of whether or not to print visual or output plotting frame}
}
\value{

}
\description{
This implementation of plot singles out a particular responder, providing the fits to the observed data as well as the probability that the "probable" model characterizes the data
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
