% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotIndividualRainbow}
\alias{plotIndividualRainbow}
\title{plotIndividualRainbow}
\usage{
plotIndividualRainbow(fittingObject, ylab0, xlab0, logAxis, yMin)
}
\arguments{
\item{fittingObject}{core frame}

\item{ylab0}{y axis label}

\item{xlab0}{x axis label}

\item{logAxis}{axis designation}

\item{yMin}{y axis lower limit}
}
\value{

}
\description{
plotIndividualRainbow
}
