% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_plot_model.R
\name{plotModelCharacterization}
\alias{plotModelCharacterization}
\title{plotModelCharacterization}
\usage{
plotModelCharacterization(fittingObject, position0, ylab0, xlab0, plotit)
}
\arguments{
\item{fittingObject}{core fitting object}

\item{position0}{(char) position of legend}

\item{ylab0}{(char) y axis label}

\item{xlab0}{(char) x axis label}

\item{plotit}{(logical) bool of whether or not to print visual or output plotting frame}
}
\value{

}
\description{
plotModelCharacterization
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
