% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotRainbowCross}
\alias{plotRainbowCross}
\title{plotRainbowCross}
\usage{
plotRainbowCross(fittingObject, metric)
}
\arguments{
\item{fittingObject}{core fitting object}

\item{metric}{...}
}
\value{

}
\description{
plotRainbowCross
}
