% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rachlin.R
\name{rachlinHyperboloidDiscountFunc}
\alias{rachlinHyperboloidDiscountFunc}
\title{Rachlin Value Function}
\usage{
rachlinHyperboloidDiscountFunc(x, lnk, s)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}

\item{s}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Rachlin Value Function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
