% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rachlinHyperboloidDiscountGradient}
\alias{rachlinHyperboloidDiscountGradient}
\title{Rachlin Gradient Helper for Nonlinear Fitting}
\usage{
rachlinHyperboloidDiscountGradient(x, lnk, s)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}

\item{s}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Rachlin Gradient Helper for Nonlinear Fitting
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
