% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{residualFunction}
\alias{residualFunction}
\title{Generalized residual call}
\usage{
residualFunction(params, x, value, valueFunction, jacobianFunction)
}
\arguments{
\item{params}{model parameters}

\item{x}{observation at point n (X)}

\item{value}{observation at point n (Y)}

\item{valueFunction}{function to get projected value}

\item{jacobianFunction}{function to create jacobian}
}
\value{
residual value of referenced function
}
\description{
General, shared method for coordinating with fitting functions
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
