% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodriguezlogue.R
\name{RodriguezLogueDiscountFunc}
\alias{RodriguezLogueDiscountFunc}
\title{Rodriguez & Logue Value Function}
\usage{
RodriguezLogueDiscountFunc(x, lnk, beta)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}

\item{beta}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Rodriguez & Logue Value Function
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
