% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodriguezlogue.R
\name{RodriguezLogueDiscountGradient}
\alias{RodriguezLogueDiscountGradient}
\title{Rodriguez & Logue Helper for Nonlinear Fitting}
\usage{
RodriguezLogueDiscountGradient(x, lnk, beta)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnk}{fitted parameter}

\item{beta}{fitted parameter}
}
\value{
projected, subjective value
}
\description{
Rodriguez & Logue Helper for Nonlinear Fitting
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>

Shawn Gilroy <sgilroy1@lsu.edu>
}
