% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{startValuesExponential}
\alias{startValuesExponential}
\title{Exponential Starting Values estimator}
\usage{
startValuesExponential(A = 1, localDat, overkill = FALSE)
}
\arguments{
\item{A}{Maximum value of commodity}

\item{localDat}{data frame with X column and Y column (0 <= Y <= 1, NOTE capitalized letters)}

\item{overkill}{optional higher resolution search}
}
\value{
A named vector of starting parameters

data frame of Screening Criteria
}
\description{
This function applies a grid search for starting parameters
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
