% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{startValuesGreenMyerson}
\alias{startValuesGreenMyerson}
\title{Green/Myerson Starting Values estimator}
\usage{
startValuesGreenMyerson(A = 1, localDat, overkill = FALSE)
}
\arguments{
\item{A}{Maximum value of commodity}

\item{localDat}{data frame with X column and Y column (0 <= Y <= 1, NOTE capitalized letters)}

\item{overkill}{optional higher resolution search}
}
\value{
A named vector of starting parameters

data frame of Screening Criteria
}
\description{
This function applies a grid search for starting parameters
}
\examples{
dat <- data.frame(X=c(1,30,180,540,1080,2160),
                 Y=c(1,0.9,0.8,0.7,0.6,0.4),
                 ids=c(1,1,1,1,1,1))
dat$Y <- dat$Y * 100
startValuesGreenMyerson(A = 100, dat, overkill = TRUE)

}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
