% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_call_dd_analyze.R
\name{dd_analyze}
\alias{dd_analyze}
\title{dd_analyze}
\usage{
dd_analyze(fittingObject, modelSelection = TRUE)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{modelSelection}{(bool) this flag determines whether or not a model selection procedure will be applied in the results frame.}
}
\value{

}
\description{
This call is the workhorse of the program. Based on the settings applied, this method applies all relevant methods and calculations to the supplied data.
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
