% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_bleichrodt.R
\name{dd_ed50_bleichrodt}
\alias{dd_ed50_bleichrodt}
\title{dd_ed50_bleichrodt}
\usage{
dd_ed50_bleichrodt(Lnk, s, b, currentData)
}
\arguments{
\item{Lnk}{parameter}

\item{s}{parameter}

\item{b}{parameter}

\item{currentData}{currentData}
}
\description{
dd_ed50_bleichrodt
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
