% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_ebertprelec.R
\name{dd_ed50_ebertprelec}
\alias{dd_ed50_ebertprelec}
\title{dd_ed50_ebertprelec}
\usage{
dd_ed50_ebertprelec(Lnk, s, currentData)
}
\arguments{
\item{Lnk}{parameter}

\item{s}{parameter}

\item{currentData}{currentData}
}
\description{
dd_ed50_ebertprelec
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
