% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{dd_ed50_exponential}
\alias{dd_ed50_exponential}
\title{dd_ed50_exponential}
\usage{
dd_ed50_exponential(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{

}
\description{
dd_ed50_exponential
}
