% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laibson.R
\name{dd_ed50_laibson}
\alias{dd_ed50_laibson}
\title{dd_ed50_laibson}
\usage{
dd_ed50_laibson(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{

}
\description{
dd_ed50_laibson
}
