% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_mazur.R
\name{dd_ed50_mazur}
\alias{dd_ed50_mazur}
\title{dd_ed50_mazur}
\usage{
dd_ed50_mazur(Lnk)
}
\arguments{
\item{Lnk}{log transformed rate parameter}
}
\description{
dd_ed50_mazur
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
