% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rachlin.R
\name{dd_ed50_rachlin}
\alias{dd_ed50_rachlin}
\title{dd_ed50_rachlin}
\usage{
dd_ed50_rachlin(Lnk, s)
}
\arguments{
\item{Lnk}{parameter}

\item{s}{parameter}
}
\description{
dd_ed50_rachlin
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
