% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodriguezlogue.R
\name{dd_fit_rodriguezlogue}
\alias{dd_fit_rodriguezlogue}
\title{dd_fit_rodriguezlogue}
\usage{
dd_fit_rodriguezlogue(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{

}
\description{
This fits a hyperbolic model to the data.
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
