% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenmyerson.R
\name{dd_mbauc_greenmyerson}
\alias{dd_mbauc_greenmyerson}
\title{dd_mbauc_greenmyerson}
\usage{
dd_mbauc_greenmyerson(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{

}
\description{
dd_mbauc_greenmyerson
}
