% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mazur.R
\name{dd_mbauc_log10_mazur}
\alias{dd_mbauc_log10_mazur}
\title{dd_mbauc_log10_mazur}
\usage{
dd_mbauc_log10_mazur(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{

}
\description{
dd_mbauc_log10_mazur
}
