% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rachlin.R
\name{dd_mbauc_log10_rachlin}
\alias{dd_mbauc_log10_rachlin}
\title{dd_mbauc_log10_rachlin}
\usage{
dd_mbauc_log10_rachlin(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{

}
\description{
dd_mbauc_log10_rachlin
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
