% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodriguezlogue.R
\name{dd_mbauc_log10_rodriguezlogue}
\alias{dd_mbauc_log10_rodriguezlogue}
\title{dd_mbauc_log10_rodriguezlogue}
\usage{
dd_mbauc_log10_rodriguezlogue(A, Lnk, b, startDelay, endDelay)
}
\arguments{
\item{A}{maximum value}

\item{Lnk}{parameter value}

\item{b}{parameter value}

\item{startDelay}{time point}

\item{endDelay}{time point}
}
\description{
dd_mbauc_log10_rodriguezlogue
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
