% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fit_mazur.R
\name{dd_mbauc_mazur}
\alias{dd_mbauc_mazur}
\title{dd_mbauc_mazur}
\usage{
dd_mbauc_mazur(A, Lnk, startDelay, endDelay)
}
\arguments{
\item{A}{maximum value}

\item{Lnk}{logged parameter value}

\item{startDelay}{time point}

\item{endDelay}{time point}
}
\description{
dd_mbauc_mazur
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
