% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mazur.R
\name{dd_mbauc_mazur}
\alias{dd_mbauc_mazur}
\title{dd_mbauc_mazur}
\usage{
dd_mbauc_mazur(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{

}
\description{
dd_mbauc_mazur
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
