% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounting.R
\name{dd_probableModel}
\alias{dd_probableModel}
\title{dd_probableModel}
\usage{
dd_probableModel(fittingObject, id)
}
\arguments{
\item{fittingObject}{core dd fitting object}

\item{id}{id tag}
}
\value{

}
\description{
Return model probabilities
}
