% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounting.R
\name{dd_screenOption}
\alias{dd_screenOption}
\title{dd_screenOption}
\usage{
dd_screenOption(
  fittingObject,
  screen,
  JB1Flag = 0.2,
  JB2Flag = 0.1,
  filterPassing = NULL
)
}
\arguments{
\item{fittingObject}{core fitting object}

\item{screen}{(bool) set screen TRUE or FALSE (i.e. NULL)}

\item{JB1Flag}{(num) bounce constant per authors (set at initial defaults)}

\item{JB2Flag}{(num) extremity change constant per authors (set at initial defaults)}

\item{filterPassing}{(char vector) which JB criteria to retain in dataset, e.g. c("JB1", "JB2")}
}
\value{

}
\description{
This call applies screening criteria to a data dataset. Specifically, it can be used to apply criteria (no filtering) or apply criteria and filter based on one or more criteria (e.g., JB1, JB2)
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
