% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounting.R
\name{dd_screenOption}
\alias{dd_screenOption}
\title{dd_screenOption}
\usage{
dd_screenOption(fittingObject, screen, filterPassing = NULL)
}
\arguments{
\item{fittingObject}{}
}
\value{

}
\description{
dd_screenOption
}
