% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleichrodt.R
\name{dd_start_bleichrodt}
\alias{dd_start_bleichrodt}
\title{dd_start_bleichrodt}
\usage{
dd_start_bleichrodt(currentData)
}
\arguments{
\item{currentData}{current data set}
}
\value{

}
\description{
Extract starting parameters
}
