% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laibson.R
\name{dd_start_laibson}
\alias{dd_start_laibson}
\title{dd_start_laibson}
\usage{
dd_start_laibson(currentData)
}
\arguments{
\item{currentData}{current data set}
}
\value{

}
\description{
Extract starting parameters
}
