% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodriguezlogue.R
\name{dd_start_rodriguezlogue}
\alias{dd_start_rodriguezlogue}
\title{dd_start_rodriguezlogue}
\usage{
dd_start_rodriguezlogue(currentData)
}
\arguments{
\item{currentData}{current data set}
}
\value{

}
\description{
Extract starting parameters
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
