% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{displayAUCFigure}
\alias{displayAUCFigure}
\title{Display most probable fitted series with AUC metric}
\usage{
displayAUCFigure(dat, results, lineWidth = 1)
}
\arguments{
\item{dat}{observed data}

\item{results}{Results of analyses for data series}

\item{lineWidth}{Line width}
}
\value{
display figure
}
\description{
This method constructs a figure that displays the most probable model
as well as the probability that it is the "true" model.  The
Model-based AUC metric is also provided for the most probable model.
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
