% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getED50}
\alias{getED50}
\title{Scoring for the log ED50}
\usage{
getED50(results)
}
\arguments{
\item{results}{Results of analyses for data series}
}
\value{
natural logarithm of the Effective Delay 50%
}
\description{
Scoring for the log ED50
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
