% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaymetrics.R
\name{getED50}
\alias{getED50}
\title{Scoring for the log ED50}
\usage{
getED50(dat, results)
}
\arguments{
\item{dat}{observed data}

\item{results}{Results of analyses for data series}
}
\value{
natural logarithm of the Effective Delay 50%
}
\description{
Methods for routing computations of the ln(ED50). When possible, exact solutions are provided and models without a straightforward exact solution have ed50 calculated numerically using a bisection search.
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
