% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areametrics.R
\name{getModelAUC}
\alias{getModelAUC}
\title{Scoring for the most probable model area}
\usage{
getModelAUC(dat, results)
}
\arguments{
\item{dat}{observed data}

\item{results}{Results of analyses for data series}
}
\value{
area beneath the fitted model
}
\description{
In this set of methods, the area beneath the fitted model is
calculated and divided by the maximum area using numerical integration
methods.  All delays are calculated in the normal scale.
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
