% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getModelAUCLog10Scaled}
\alias{getModelAUCLog10Scaled}
\title{Scoring for the most probable model area, in log10 space}
\usage{
getModelAUCLog10Scaled(dat, results)
}
\arguments{
\item{dat}{observed data}

\item{results}{Results of analyses for data series}
}
\value{
area beneath the fitted model, in log10 space
}
\description{
Scoring for the most probable model area, in log10 space
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
