% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{integrandEbertPrelec}
\alias{integrandEbertPrelec}
\title{Ebert & Prelec's CS Integrand helper}
\usage{
integrandEbertPrelec(x, lnK, s)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnK}{fitted parameter}

\item{s}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
Ebert & Prelec's CS Integrand helper
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
