% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrandHyp}
\alias{integrandHyp}
\title{Hyperbolic Integrand helper}
\usage{
integrandHyp(x, lnK)
}
\arguments{
\item{x}{observation at point n (X)}

\item{lnK}{fitted parameter}
}
\value{
Numerical Integration Projection
}
\description{
This integrand helper is a projection of the integrand with delays represented as normal
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
