% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounting.R
\name{jacobianMatrix}
\alias{jacobianMatrix}
\title{Generalized Jacobian call}
\usage{
jacobianMatrix(params, x, value, valueFunction, jacobianFunction)
}
\arguments{
\item{params}{model parameters}

\item{x}{observation at point n (X)}

\item{value}{observation at point n (Y)}

\item{valueFunction}{function to get projected value}

\item{jacobianFunction}{function to create jacobian}
}
\value{
difference value for jacobian
}
\description{
General, shared method for constructing the Jacobian matrix. Routes a supplied "jacobianFunction" with pre-computed derivatives to construct matrix with observed data and supplied parameters.
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
