% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{johnsonBickelScreen}
\alias{johnsonBickelScreen}
\title{Perform Johnson & Bickel Screen}
\usage{
johnsonBickelScreen(fittingObject)
}
\arguments{
\item{fittingObject}{core fitting object}
}
\value{
A data frame of model screenings


}
\description{
This function applies the Johnson & Bickel screening criteria to included data series. The result of this procedure is a TRUE/FALSE response to one of two screening criteria.
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
