% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{johnsonBickelScreen}
\alias{johnsonBickelScreen}
\title{Perform Johnson & Bickel Screen}
\usage{
johnsonBickelScreen(dat, idCol = "id")
}
\arguments{
\item{dat}{data frame with X column and Y column (0 <= Y <= 1)}

\item{idCol}{id column}
}
\value{
A data frame of model screenings

data frame of Screening Criteria
}
\description{
This function applies the Johnson & Bickel screening criteria to included data series
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
