% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{johnsonBickelScreen}
\alias{johnsonBickelScreen}
\title{Perform Johnson & Bickel Screen}
\usage{
johnsonBickelScreen(dat, idCol = "id")
}
\arguments{
\item{dat}{data frame with X column and Y column (0 <= Y <= 1, NOTE capitalized letters)}

\item{idCol}{string content identifying participants (e.g., idCol = "ids")}
}
\value{
A data frame of model screenings

data frame of Screening Criteria
}
\description{
This function applies the Johnson & Bickel screening criteria to included data series. The result of this procedure is a TRUE/FALSE response to one of two screening criteria. These are included by default in all model selection calls.
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
