% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logLik.nls.lm}
\alias{logLik.nls.lm}
\title{minpack.lm logLik hack}
\usage{
\method{logLik}{nls.lm}(fit, REML = FALSE, ...)
}
\arguments{
\item{fit}{nls.lm fitted model}

\item{REML}{determine whether or not to use ML (FALSE by default)}

\item{...}{inherit other args as necessary}
}
\value{
provide a logLik class for AIC/BIC
}
\description{
This function constructs a class, derived from an nls.lm object, similar to that of the logLik function in nls. This allows for native calls of the AIC and BIC functions from stats, using nls.lm fit objects.
}
\author{
Katharine Mullen <kate@few.vu.nl>
}
