% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_plot.R
\name{plot.discountingtools}
\alias{plot.discountingtools}
\title{plot.discountingtools}
\usage{
\method{plot}{discountingtools}(
  fittingObject,
  which = "ind",
  position0 = "bottomleft",
  ylab0 = "Subjective Value",
  xlab0 = "Delay",
  logAxis = "x",
  yMin = 0.01,
  id = NULL,
  plotit = TRUE
)
}
\arguments{
\item{fittingObject}{core fitting object}

\item{which}{(char) type of plot to show, based on fits}

\item{position0}{(char) position of legend}

\item{ylab0}{(char) y axis label}

\item{xlab0}{(char) x axis label}

\item{logAxis}{(char) axis designation}

\item{yMin}{(num) y axis lower limit}

\item{id}{(num) participant number to focus}

\item{plotit}{(logical) bool of whether or not to print visual or output plotting frame}
}
\value{

}
\description{
This method overrides the base plot function to provide various plots relevant to the user.
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
