% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotRainbowCross}
\alias{plotRainbowCross}
\title{plotRainbowCross}
\usage{
plotRainbowCross(fittingObject, metric, plotit)
}
\arguments{
\item{fittingObject}{core fitting object}

\item{metric}{(char) the cross model metric to be displayed}

\item{plotit}{(logical) bool of whether or not to print visual or output plotting frame}
}
\value{

}
\description{
plotRainbowCross
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
