% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{residualFunction}
\alias{residualFunction}
\title{Generalized residual call}
\usage{
residualFunction(params, x, value, valueFunction, jacobianFunction)
}
\arguments{
\item{params}{model parameters}

\item{x}{observation at point n (X)}

\item{value}{observation at point n (Y)}

\item{valueFunction}{function to get projected value}

\item{jacobianFunction}{function to create jacobian}
}
\value{
residual value of referenced function
}
\description{
General, shared method for coordinating nls.lm fitting calls. Routes a supplied "valueFunction" with observed data and supplied parameters.
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
