% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.discountingtools}
\alias{summary.discountingtools}
\title{summary.discountingtools}
\usage{
\method{summary}{discountingtools}(fittingObject)
}
\arguments{
\item{fittingObject}{}
}
\value{

}
\description{
Override summary output
}
