% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.discountingtools}
\alias{summary.discountingtools}
\title{summary.discountingtools}
\usage{
\method{summary}{discountingtools}(fittingObject, detailed = TRUE)
}
\arguments{
\item{fittingObject}{core fitting object}

\item{detailed}{enable additional model metrics (default TRUE)}
}
\value{

}
\description{
Override summary output. Rather than display the core fitting object, a data frame block of results is provided to the user for easy interpretation and further analysis
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
